feat: support passwordless Target database accounts (#15)
* feat: allow explicit empty Target passwords

* feat: encrypt passwordless Target credentials

* feat: resolve explicit empty Target passwords

* test: cover passwordless credential encryption

* test: add passwordless Target credential coverage

* feat: support passwordless Targets in UI

* test: cover passwordless Target request models

* fix: serialize request validation errors safely

* fix: avoid reserved MySQL probe alias

* test: cover legacy MySQL probe alias compatibility

* fix: store large MySQL column lengths as bigint

* fix: migrate catalog column length to bigint

* test: cover large catalog column lengths

+++ backend/app/models/catalog.py
@@ -95,7 +95,7 @@ class CatalogColumn(CatalogBase):
     ordinal_position: Mapped[int] = mapped_column(Integer, nullable=False)
     column_name: Mapped[str] = mapped_column(String(200), nullable=False)
     data_type: Mapped[str] = mapped_column(String(100), nullable=False)
-    character_maximum_length: Mapped[int | None] = mapped_column(Integer)
+    character_maximum_length: Mapped[int | None] = mapped_column(BigInteger)
     numeric_precision: Mapped[int | None] = mapped_column(Integer)
     numeric_scale: Mapped[int | None] = mapped_column(Integer)
     is_nullable: Mapped[bool] = mapped_column(Boolean, nullable=False, default=True)
